hs12-flask-api-raamatute-otsing/requirements.txt
Flask
requests
azure-storage-blob
flask-cors
python-dotenv
